+++ ec/modules/dispatch.py
@@ -6,20 +6,13 @@ def init(argv):
   flag = argv.pop(0) if argv[0][0] == '-' else None
   
   if flag == '-h':
-    return show_help()
+    print get_help_text()
   
   try:
     execCommand(argv, flag == '-p')
     
   except HandledException as e:
     err(e, 1)
-    
-def show_help(is_error=0):
-  if is_error:
-    err(get_help_text(), 1)
-    
-  else:
-    print get_help_text()
     
 def get_help_text():
   from helper_tasks import listMemberHelps
@@ -34,3 +27,4 @@ def get_help_text():
   ) + '  ' + '\n  '.join([('%s  %s' % (name, desc))[:60] for name, desc in listMemberHelps('')])
   
   return text
+