+++ scripts/examples/simple.py
@@ -1,4 +1,4 @@
-from ec.ec import start, task, arg, group
+from ec.config import start, task, arg, group
 
 @task
 @arg('arg1', type=int, desc= 'Value for arg1')